EDA and Linear regression model
Fyntra is the largest online clothing company in USA. It sells clothing online,but they also have in-store style and clothing advice sessions. Customers come into the store, have sessions/meetings with a personal stylist, then can go home and order either on a mobile app or website for the clothes they want.Company wants to decidewhether to focus the effort on mobile app experience or its website. As a drastic measure it is also evaluating to shut down the website. They need a ML model which will help them make the right decision.